create version of llm instructions with and without analysis
